[apps] Solve duplicate function error during benchmark
