+++ apps/benchmarks/main.c
@@ -70,6 +70,9 @@
 #elif defined(PATHFINDER)
 #include "benchmark/pathfinder.bmark"
 
+#elif defined(ROI_ALIGN)
+#include "benchmark/roi_align.bmark"
+
 #else
 #error                                                                         \
     "Error, no kernel was specified. Please, run 'make bin/benchmarks ENV_DEFINES=-D${KERNEL}', where KERNEL contains the kernel to benchmark. For example: 'make bin/benchmarks ENV_DEFINES=-DIMATMUL'."
